requirements.txt
langchain
langchainhub
langchain-openai
langchain-pinecone
langchain-community
pinecone-client
python-dotenv
ipykernel
streamlit